refactor: query client, services, hooks
