--- src/features/authentication/hooks/useLoginForm.ts
@@ -1,32 +1,0 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { message } from "antd";
-
-import { authService } from "@/services/auth";
-import { LoginFormValues } from "@/features/authentication/types/login";
-
-export const useLoginForm = () => {
-  const navigate = useNavigate();
-  const [loading, setLoading] = useState(false);
-  const login = authService.useLogin();
-
-  const handleLogin = async (values: LoginFormValues) => {
-    try {
-      setLoading(true);
-      await login.mutateAsync(values);
-      message.success("Logged in successfully!");
-      navigate("/dashboard");
-    } catch (error: unknown) {
-      message.error(
-        error instanceof Error ? error.message : "Invalid credentials"
-      );
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return {
-    loading,
-    handleLogin,
-  };
-};
