+++ src/services/auth.ts
@@ -1,5 +1,4 @@
 import { useMutation } from "@tanstack/react-query";
-
 import { useAuthStore } from "@/stores/useAuthStore";
 import { axiosInstance } from "./api";
 
@@ -8,26 +7,24 @@ interface LoginCredentials {
   password: string;
 }
 
-export const authService = {
-  // Login mutation
-  useLogin() {
-    return useMutation({
-      mutationFn: async (credentials: LoginCredentials) => {
-        const response = await axiosInstance.post("auth/login", credentials);
-        const { token, user, permissions } = response.data;
+export const useLogin = () => {
+  return useMutation({
+    mutationFn: async (credentials: LoginCredentials) => {
+      const response = await axiosInstance.post("auth/login", credentials);
+      const { token, user, permissions } = response.data;
 
-        // Update auth store
-        useAuthStore.getState().setToken(token);
-        useAuthStore.getState().setUser(user);
-        useAuthStore.getState().setPermissions(permissions);
+      // Update auth store
+      useAuthStore.getState().setToken(token);
+      useAuthStore.getState().setUser(user);
+      useAuthStore.getState().setPermissions(permissions);
 
-        return response.data;
-      },
-    });
-  },
+      return response.data;
+    },
+  });
+};
 
-  // Logout mutation
-  useLogout() {
+export const useLogout = () => {
+  return () => {
     // Clear local state immediately
     useAuthStore.getState().clear();
     // Optionally notify backend (fire-and-forget)
@@ -35,7 +32,5 @@ export const authService = {
       // Ignore errors since we're already logged out locally
       console.debug("Backend logout notification failed");
     });
-  },
+  };
 };
-
-export default authService;
